Jetpack: Deprecate `jetpack_require_lib` (#25886)
The `jetpack_require_lib` function exists for compatibility with the way
Jetpack code is shared with WordPress.com Simple sites (i.e. Fusion).

We're now changing the way that the code sharing works in a way that
means this function will no longer be needed and would need extra logic
to make work in a backwards- and forwards-compatible manner. Instead
we're going to replace all uses of the function with direct requiring of
the relevant files, then remove it.

Note this PR does not remove uses (that needs to be done as part of the
de-Fusioning of each feature), and so does not add `_deprecated_function`
and `apply_filters_deprecated` calls yet. That will come once we've
removed all uses in Jetpack (and WordPress.com). Then we'll wait a few
Jetpack releases after that before removing the function entirely.

Internal reference: p9dueE-5EZ-p2

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2950894587