VideoPress: Allow empty video description and caption (#26564)
* Fix caption and description checks to allow empty values

Now, when an empty value is sent, it will update the data with the empty string. The fields will only be ignored when they are not present on the request, so it's still safe to do requests using only some of the fields.

* Add changelog file

Co-authored-by: Damián Suárez <rdsuarez@gmail.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3176057680