VideoPress: add endpoint to fetch site data without relying on active Jetpack plugin (#26652)
* Add class to register videopress/v1/site endpoint

The endpoint is READABLE and returns the same data returned by jetpack/v4/site, but without requiring an active Jetpack connection.

* Enable the new endpoint by initing it's class

* Add proper class to deal with the site data

The Site class is the provider of all the data that will be returned on the new videopress/v1/site, as well as any other use that we may find relevant for it.

* Refactor the rest endpoint class to use the new Site class

Moved all the logic related to requesting site data to the new Site class, so the VideoPress_Rest_Api_V1_Site can simply use it to fetch the data and return it as a response.

* Add changelog file

* Create a central method on Data class to provide the Jetpack blog ID

* Refactor the Site class to rely on the new Data blog ID provider

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3198171914