VideoPress block: Extract poster-related functionality from useSyncMedia hook (#29973)
Refactor part of preparation for implementing a native version of the useSyncMedia hook. Note, this is the second of a series of planned changes for splitting the hook's functionality into smaller hooks, with the goal of making it more straightforward to implement the native functionality. The first of these changes can be found at #29898.

See wordpress-mobile/gutenberg-mobile#5611 for more context.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4640115046