Feature Hints: add sharing block to the list of suggestions (#35577)
Feature Hints suggest the legacy sharing module whenever someone looks for sharing buttons in the plugin search. It is useful, but on block-based themes we can now suggest a better solution: the new sharing buttons block.
This adds logic so we suggest a different solution based on the theme in use on your site.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7870105350