ci: Check that React-using projects use react eslint rules (#39214)
I noticed that several of our projects are using React but did not have
our `jetpack-js-tools/eslintrc/react` eslint ruleset enabled. This adds
a linting check for that, and enables it on most of the projects where
it was not enabled.

For a few I instead added a `@todo` comment about turning it on, due to
a lot of cleanup being needed in order to successfully do so. I hope
people more familiar with React and these projects will follow up on
that.

* js-packages/ai-client:
   32 × react/jsx-no-bind
   33 × react-hooks/exhaustive-deps
* packages/videopress:
    8 × react-hooks/rules-of-hooks
   69 × react-hooks/exhaustive-deps
  157 × react/jsx-no-bind
* plugins/boost:
    1 × react/no-danger
   81 × react/jsx-no-bind
* plugins/crm:
    1 × react-hooks/rules-of-hooks
    2 × react/jsx-no-bind
    3 × react-hooks/exhaustive-deps

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/10722474820

Upstream-Ref: Automattic/jetpack@d586800b8457f6106c1fecede9abf95d13ccfd24